Prevent SSL test script from failing

+++ ssl-test-setup/create-key-store.sh
@@ -6,7 +6,7 @@ HOSTNAME="$1"
 PASSWORD="$2"
 INSTALL_DIR="$3"
 
-rm -r "$INSTALL_DIR" || true
+[[ -d "$INSTALL_DIR" ]] && rm -r "$INSTALL_DIR" || true
 mkdir -p "$INSTALL_DIR"
 
 cp "./ssl-test-setup/req.conf" "$INSTALL_DIR/req.conf"
